chore: move components out of shadcn folder (#22)
